requirements.txt
langchain
openai
pyyaml
streamlit
unstructured
pypandoc
pdflatex
pdf2image
# streamlit-audiorec
# assemblyai 
# assemblyai[extras]
PyPDF2
python-dotenv
faiss-cpu
streamlit-extras
tiktoken
sentence-transformers
pdfminer.six